task1/src/requirements.txt
python-dotenv
python-telegram-bot
requests
boto3